+++ examples/Redis.OM.FullTextSearch/Models/Awards.cs
@@ -2,7 +2,7 @@
 
 using Redis.OM.Modeling;
 
-namespace Redis.OM.Full_Text_Search_within_an_Index.Models
+namespace Redis.OM.FullTextSearch.Models
 {
     [Document(IndexName = "awards-idx", StorageType = StorageType.Json)]
     public class Awards
