+++ examples/Redis.OM.FullTextSearch/Redis.OM.FullTextSearch.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32901.215
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Redis.OM.Full-Text_Search_within_an_Index", "Redis.OM.Full-Text_Search_within_an_Index.csproj", "{5F2B8BBC-EB2D-40AC-928E-20DEB3EF873E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Redis.OM.FullTextSearch", "Redis.OM.FullTextSearch.csproj", "{5F2B8BBC-EB2D-40AC-928E-20DEB3EF873E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
